requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.10.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
email-validator==2.2.0
